@@ -17,10 +17,10 @@
         }
     ],
     "require": {
-        "php": ">=7.0.0"
+        "php": ">=8.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "~5.7 >=5.7.11",
+        "phpunit/phpunit": "~9.6",
         "knplabs/gaufrette": "~0.5",
         "symfony/yaml": "~3.4"
     },
