@@ -58,9 +58,9 @@ class PathFactory extends AbstractPathFactory
     /**
      * Creates a path factory
      *
-     * @param   string              $protocol    The protocol (such as "git")
-     * @param   Binary|string|null  $git         The Git binary
-     * @param   RepositoryRegistry  $map         The repository registry
+     * @param   string                   $protocol    The protocol (such as "git")
+     * @param   Binary|string|null       $git         The Git binary
+     * @param   RepositoryRegistry|null  $map         The repository registry
      */
     public function __construct($protocol, $git = null, RepositoryRegistry $map = null)
     {
