@@ -23,18 +23,19 @@
 
 namespace TQ\Tests\Git\StreamWrapper;
 
+use PHPUnit\Framework\TestCase;
 use TQ\Git\Cli\Binary;
 use TQ\Git\Repository\Repository;
 use TQ\Git\StreamWrapper\StreamWrapper;
 use TQ\Tests\Helper;
 
-class DirectoryTest extends \PHPUnit_Framework_TestCase
+class DirectoryTest extends TestCase
 {
     /**
      * Sets up the fixture, for example, open a network connection.
      * This method is called before a test is executed.
      */
-    protected function setUp()
+    protected function setUp(): void
     {
         Helper::removeDirectory(TESTS_TMP_PATH);
         Helper::createDirectory(TESTS_TMP_PATH);
@@ -74,7 +75,7 @@ class DirectoryTest extends \PHPUnit_Framework_TestCase
      * Tears down the fixture, for example, close a network connection.
      * This method is called after a test is executed.
      */
-    protected function tearDown()
+    protected function tearDown(): void
     {
         Helper::removeDirectory(TESTS_TMP_PATH);
 
@@ -85,7 +86,7 @@ class DirectoryTest extends \PHPUnit_Framework_TestCase
      *
      * @return  Repository
      */
-    protected function getRepository()
+    protected function getRepository(): Repository
     {
         return Repository::open(TESTS_REPO_PATH_1, new Binary(GIT_BINARY));
     }
@@ -481,7 +482,7 @@ class DirectoryTest extends \PHPUnit_Framework_TestCase
         $this->assertStringStartsWith('dir', $missingDirectoryName);
 
         $missingDirectoryContents = $c->listDirectory($missingDirectoryName, 'HEAD');
-        $this->assertInternalType('array', $missingDirectoryContents);
+        $this->assertIsArray($missingDirectoryContents);
         $this->assertEquals('file.txt', current($missingDirectoryContents));
 
         $c->removeFile($missingDirectoryName, null, true);
@@ -489,4 +490,3 @@ class DirectoryTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals($c->listDirectory($missingDirectoryName, 'HEAD^'), $missingDirectoryContents);
     }
 }
-
