@@ -23,18 +23,19 @@
 
 namespace TQ\Tests\Svn\Repository;
 
+use PHPUnit\Framework\TestCase;
 use TQ\Svn\Cli\Binary;
 use TQ\Svn\Repository\Repository;
 use TQ\Tests\Helper;
 use TQ\Vcs\Repository\Transaction;
 
-class ModificationTest extends \PHPUnit_Framework_TestCase
+class ModificationTest extends TestCase
 {
     /**
      * Sets up the fixture, for example, open a network connection.
      * This method is called before a test is executed.
      */
-    protected function setUp()
+    protected function setUp(): void
     {
         Helper::removeDirectory(TESTS_TMP_PATH);
         Helper::createDirectory(TESTS_TMP_PATH);
@@ -61,7 +62,7 @@ class ModificationTest extends \PHPUnit_Framework_TestCase
      * Tears down the fixture, for example, close a network connection.
      * This method is called after a test is executed.
      */
-    protected function tearDown()
+    protected function tearDown(): void
     {
         Helper::removeDirectory(TESTS_TMP_PATH);
     }
@@ -70,7 +71,7 @@ class ModificationTest extends \PHPUnit_Framework_TestCase
      *
      * @return  Repository
      */
-    protected function getRepository()
+    protected function getRepository(): Repository
     {
         return Repository::open(TESTS_REPO_PATH_1, new Binary(SVN_BINARY));
     }
@@ -85,7 +86,7 @@ class ModificationTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals('Test', file_get_contents(TESTS_REPO_PATH_1.'/test.txt'));
 
         $commit = $c->showCommit($revision);
-        $this->assertContains('A /test.txt', $commit);
+        $this->assertStringContainsString('A /test.txt', $commit);
     }
 
     public function testAddFileInSubdirectory()
@@ -98,7 +99,7 @@ class ModificationTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals('Test', file_get_contents(TESTS_REPO_PATH_1.'/directory/test.txt'));
 
         $commit = $c->showCommit($revision);
-        $this->assertContains('A /directory/test.txt', $commit);
+        $this->assertStringContainsString('A /directory/test.txt', $commit);
     }
 
     public function testAddFileInSecondLevelSubdirectory()
@@ -111,7 +112,7 @@ class ModificationTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals('Test', file_get_contents(TESTS_REPO_PATH_1.'/dirA/dirB/test.txt'));
 
         $commit = $c->showCommit($revision);
-        $this->assertContains('A /dirA/dirB/test.txt', $commit);
+        $this->assertStringContainsString('A /dirA/dirB/test.txt', $commit);
     }
 
     public function testAddMultipleFiles()
@@ -121,7 +122,7 @@ class ModificationTest extends \PHPUnit_Framework_TestCase
             $revision   = $c->writeFile(sprintf('test_%s.txt', $i), $i);
             $this->assertEquals($i + 2, $revision);
             $commit = $c->showCommit($revision);
-            $this->assertContains(sprintf('A /test_%d.txt', $i), $commit);
+            $this->assertStringContainsString(sprintf('A /test_%d.txt', $i), $commit);
         }
 
         for ($i = 0; $i < 5; $i++) {
@@ -136,10 +137,10 @@ class ModificationTest extends \PHPUnit_Framework_TestCase
         $revision   = $c->removeFile('file_0.txt');
         $this->assertEquals(2, $revision);
 
-        $this->assertFileNotExists(TESTS_REPO_PATH_1.'/file_0.txt');
+        $this->assertFileDoesNotExist(TESTS_REPO_PATH_1.'/file_0.txt');
 
         $commit = $c->showCommit($revision);
-        $this->assertContains('D /file_0.txt', $commit);
+        $this->assertStringContainsString('D /file_0.txt', $commit);
     }
 
     public function testRemoveMultipleFiles()
@@ -149,11 +150,11 @@ class ModificationTest extends \PHPUnit_Framework_TestCase
             $revision   = $c->removeFile(sprintf('file_%s.txt', $i), $i);
             $this->assertEquals($i + 2, $revision);
             $commit = $c->showCommit($revision);
-            $this->assertContains(sprintf('D /file_%d.txt', $i), $commit);
+            $this->assertStringContainsString(sprintf('D /file_%d.txt', $i), $commit);
         }
 
         for ($i = 0; $i < 5; $i++) {
-            $this->assertFileNotExists(TESTS_REPO_PATH_1.sprintf('/file_%d.txt', $i));
+            $this->assertFileDoesNotExist(TESTS_REPO_PATH_1.sprintf('/file_%d.txt', $i));
         }
     }
 
@@ -164,15 +165,15 @@ class ModificationTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals(2, $revision);
 
         for ($i = 0; $i < 5; $i++) {
-            $this->assertFileNotExists(TESTS_REPO_PATH_1.sprintf('/file_%d.txt', $i));
+            $this->assertFileDoesNotExist(TESTS_REPO_PATH_1.sprintf('/file_%d.txt', $i));
         }
 
         $commit = $c->showCommit($revision);
-        $this->assertContains('D /file_0.txt', $commit);
-        $this->assertContains('D /file_1.txt', $commit);
-        $this->assertContains('D /file_2.txt', $commit);
-        $this->assertContains('D /file_3.txt', $commit);
-        $this->assertContains('D /file_4.txt', $commit);
+        $this->assertStringContainsString('D /file_0.txt', $commit);
+        $this->assertStringContainsString('D /file_1.txt', $commit);
+        $this->assertStringContainsString('D /file_2.txt', $commit);
+        $this->assertStringContainsString('D /file_3.txt', $commit);
+        $this->assertStringContainsString('D /file_4.txt', $commit);
     }
 
     public function testRemoveSubdirectory()
@@ -183,10 +184,10 @@ class ModificationTest extends \PHPUnit_Framework_TestCase
         $revision   = $c->removeFile('subdirectory', null, true);
         $this->assertEquals(3, $revision);
 
-        $this->assertFileNotExists(TESTS_REPO_PATH_1.'/subdirectory');
+        $this->assertFileDoesNotExist(TESTS_REPO_PATH_1.'/subdirectory');
 
         $commit = $c->showCommit($revision);
-        $this->assertContains('deleted file', $commit);
+        $this->assertStringContainsString('deleted file', $commit);
     }
 
     public function testMoveFile()
@@ -195,12 +196,12 @@ class ModificationTest extends \PHPUnit_Framework_TestCase
         $revision   = $c->renameFile('file_0.txt', 'test.txt');
         $this->assertEquals(2, $revision);
 
-        $this->assertFileNotExists(TESTS_REPO_PATH_1.'/file_0.txt');
+        $this->assertFileDoesNotExist(TESTS_REPO_PATH_1.'/file_0.txt');
         $this->assertFileExists(TESTS_REPO_PATH_1.'/test.txt');
 
         $commit = $c->showCommit($revision);
-        $this->assertContains('A /test.txt', $commit);
-        $this->assertContains('/file_0.txt', $commit);
+        $this->assertStringContainsString('A /test.txt', $commit);
+        $this->assertStringContainsString('/file_0.txt', $commit);
     }
 
     public function testReset()
@@ -213,13 +214,13 @@ class ModificationTest extends \PHPUnit_Framework_TestCase
         $this->assertTrue($c->isDirty());
         $c->reset();
         $this->assertFalse($c->isDirty());
-        $this->assertFileNotExists($file);
+        $this->assertFileDoesNotExist($file);
 
         file_put_contents($file, 'Test');
         $c->add(array('test.txt'));
         $c->reset();
         $this->assertFalse($c->isDirty());
-        $this->assertFileNotExists($file);
+        $this->assertFileDoesNotExist($file);
     }
 
     public function testTransactionalChangesNoException()
@@ -252,12 +253,12 @@ class ModificationTest extends \PHPUnit_Framework_TestCase
         $this->assertContains('test_4.txt', $list);
 
         $commit = $c->showCommit($result->getCommitHash());
-        $this->assertContains($result->getCommitMsg(), $commit);
-        $this->assertContains('A /test_0.txt', $commit);
-        $this->assertContains('A /test_1.txt', $commit);
-        $this->assertContains('A /test_2.txt', $commit);
-        $this->assertContains('A /test_3.txt', $commit);
-        $this->assertContains('A /test_4.txt', $commit);
+        $this->assertStringContainsString($result->getCommitMsg(), $commit);
+        $this->assertStringContainsString('A /test_0.txt', $commit);
+        $this->assertStringContainsString('A /test_1.txt', $commit);
+        $this->assertStringContainsString('A /test_2.txt', $commit);
+        $this->assertStringContainsString('A /test_3.txt', $commit);
+        $this->assertStringContainsString('A /test_4.txt', $commit);
     }
 
     public function testTransactionalChangesException()
@@ -304,10 +305,10 @@ class ModificationTest extends \PHPUnit_Framework_TestCase
         $this->assertContains('file_4.txt', $list);
 
         $commit = $c->showCommit($result->getCommitHash());
-        $this->assertContains($result->getCommitMsg(), $commit);
-        $this->assertContains('D /file_0.txt', $commit);
-        $this->assertContains('D /file_1.txt', $commit);
-        $this->assertContains('A /test.txt', $commit);
+        $this->assertStringContainsString($result->getCommitMsg(), $commit);
+        $this->assertStringContainsString('D /file_0.txt', $commit);
+        $this->assertStringContainsString('D /file_1.txt', $commit);
+        $this->assertStringContainsString('A /test.txt', $commit);
     }
 
     public function testTransactionalNoChanges()
@@ -325,4 +326,3 @@ class ModificationTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals($currentCommit, $c->getCurrentCommit());
     }
 }
-
