@@ -23,18 +23,19 @@
 
 namespace TQ\Tests\Svn\StreamWrapper;
 
+use PHPUnit\Framework\TestCase;
 use TQ\Svn\Cli\Binary;
 use TQ\Svn\Repository\Repository;
 use TQ\Svn\StreamWrapper\StreamWrapper;
 use TQ\Tests\Helper;
 
-class FileWriteTest extends \PHPUnit_Framework_TestCase
+class FileWriteTest extends TestCase
 {
     /**
      * Sets up the fixture, for example, open a network connection.
      * This method is called before a test is executed.
      */
-    protected function setUp()
+    protected function setUp(): void
     {
         Helper::removeDirectory(TESTS_TMP_PATH);
         Helper::createDirectory(TESTS_TMP_PATH);
@@ -63,7 +64,7 @@ class FileWriteTest extends \PHPUnit_Framework_TestCase
      * Tears down the fixture, for example, close a network connection.
      * This method is called after a test is executed.
      */
-    protected function tearDown()
+    protected function tearDown(): void
     {
         Helper::removeDirectory(TESTS_TMP_PATH);
 
@@ -74,7 +75,7 @@ class FileWriteTest extends \PHPUnit_Framework_TestCase
      *
      * @return  Repository
      */
-    protected function getRepository()
+    protected function getRepository(): Repository
     {
         return Repository::open(TESTS_REPO_PATH_1, new Binary(SVN_BINARY));
     }
@@ -91,7 +92,7 @@ class FileWriteTest extends \PHPUnit_Framework_TestCase
 
         $c      = $this->getRepository();
         $commit = $c->showCommit($c->getCurrentCommit());
-        $this->assertRegExp('~^\s{3}A /test.txt$~m', $commit);
+        $this->assertMatchesRegularExpression('~^\s{3}A /test.txt$~m', $commit);
     }
 
     public function testWriteNewFileWithContext()
@@ -112,9 +113,9 @@ class FileWriteTest extends \PHPUnit_Framework_TestCase
 
         $c      = $this->getRepository();
         $commit = $c->showCommit($c->getCurrentCommit());
-        $this->assertRegExp('~^\s{3}A /test.txt$~m', $commit);
-        $this->assertContains('Hello World', $commit);
-        $this->assertContains('Luke Skywalker <skywalker@deathstar.com>', $commit);
+        $this->assertMatchesRegularExpression('~^\s{3}A /test.txt$~m', $commit);
+        $this->assertStringContainsString('Hello World', $commit);
+        $this->assertStringContainsString('Luke Skywalker <skywalker@deathstar.com>', $commit);
     }
 
     public function testWriteExistingFile()
@@ -129,7 +130,7 @@ class FileWriteTest extends \PHPUnit_Framework_TestCase
 
         $c      = $this->getRepository();
         $commit = $c->showCommit($c->getCurrentCommit());
-        $this->assertRegExp('~^\s{3}M /file_0.txt$~m', $commit);
+        $this->assertMatchesRegularExpression('~^\s{3}M /file_0.txt$~m', $commit);
     }
 
     public function testAppendExistingFile()
@@ -144,7 +145,7 @@ class FileWriteTest extends \PHPUnit_Framework_TestCase
 
         $c      = $this->getRepository();
         $commit = $c->showCommit($c->getCurrentCommit());
-        $this->assertRegExp('~^\s{3}M /file_0.txt$~m', $commit);
+        $this->assertMatchesRegularExpression('~^\s{3}M /file_0.txt$~m', $commit);
     }
 
     public function testWriteNewFileWithX()
@@ -159,17 +160,7 @@ class FileWriteTest extends \PHPUnit_Framework_TestCase
 
         $c      = $this->getRepository();
         $commit = $c->showCommit($c->getCurrentCommit());
-        $this->assertRegExp('~^\s{3}A /test.txt$~m', $commit);
-    }
-
-    /**
-     * @expectedException \PHPUnit_Framework_Error_Warning
-     */
-    public function testWriteExistingFileWithXFails()
-    {
-        $filePath   = sprintf('svn://%s/file_0.txt', TESTS_REPO_PATH_1);
-        $file       = fopen($filePath, 'x');
-        fwrite($file, 'Test');
+        $this->assertMatchesRegularExpression('~^\s{3}A /test.txt$~m', $commit);
     }
 
     public function testWriteNewFileWithC()
@@ -184,7 +175,7 @@ class FileWriteTest extends \PHPUnit_Framework_TestCase
 
         $c      = $this->getRepository();
         $commit = $c->showCommit($c->getCurrentCommit());
-        $this->assertRegExp('~^\s{3}A /test.txt$~m', $commit);
+        $this->assertMatchesRegularExpression('~^\s{3}A /test.txt$~m', $commit);
     }
 
     public function testWriteExistingFileWithC()
@@ -200,7 +191,7 @@ class FileWriteTest extends \PHPUnit_Framework_TestCase
 
         $c      = $this->getRepository();
         $commit = $c->showCommit($c->getCurrentCommit());
-        $this->assertRegExp('~^\s{3}M /file_0.txt$~m', $commit);
+        $this->assertMatchesRegularExpression('~^\s{3}M /file_0.txt$~m', $commit);
     }
 
     public function testWriteAndReadNewFile()
@@ -219,7 +210,7 @@ class FileWriteTest extends \PHPUnit_Framework_TestCase
 
         $c      = $this->getRepository();
         $commit = $c->showCommit($c->getCurrentCommit());
-        $this->assertRegExp('~^\s{3}A /test.txt$~m', $commit);
+        $this->assertMatchesRegularExpression('~^\s{3}A /test.txt$~m', $commit);
     }
 
     public function testWriteNewFileWithFilePutContents()
@@ -232,7 +223,7 @@ class FileWriteTest extends \PHPUnit_Framework_TestCase
 
         $c      = $this->getRepository();
         $commit = $c->showCommit($c->getCurrentCommit());
-        $this->assertRegExp('~^\s{3}A /test.txt$~m', $commit);
+        $this->assertMatchesRegularExpression('~^\s{3}A /test.txt$~m', $commit);
     }
 
     public function testWriteExistingFileWithFilePutContents()
@@ -245,7 +236,7 @@ class FileWriteTest extends \PHPUnit_Framework_TestCase
 
         $c      = $this->getRepository();
         $commit = $c->showCommit($c->getCurrentCommit());
-        $this->assertRegExp('~^\s{3}M /file_0.txt$~m', $commit);
+        $this->assertMatchesRegularExpression('~^\s{3}M /file_0.txt$~m', $commit);
     }
 
     public function testAppendExistingFileWithFilePutContents()
@@ -258,7 +249,6 @@ class FileWriteTest extends \PHPUnit_Framework_TestCase
 
         $c      = $this->getRepository();
         $commit = $c->showCommit($c->getCurrentCommit());
-        $this->assertRegExp('~^\s{3}M /file_0.txt$~m', $commit);
+        $this->assertMatchesRegularExpression('~^\s{3}M /file_0.txt$~m', $commit);
     }
 }
-
